refactor(core): consolidate forward declaration files into single forward.h
Merge three separate forward declaration files into a single authoritative
file to comply with Simple Design principle of "No Duplication".

Changes:
- Merge core/forward_declarations.h and impl/typed_pool/forward_declarations.h
  into the main forward.h file
- Add missing type declarations: callback_typed_job_t, typed_thread_pool_builder,
  typed_job_interface, and job_types with uint8_t base type
- Update internal includes to use the consolidated forward.h
- Remove duplicate forward declaration files
- Update CMakeLists.txt to remove deleted file reference

Closes #353
