--- include/kcenon/thread/core/forward_declarations.h
@@ -1,70 +1,0 @@
-#pragma once
-
-/*****************************************************************************
-BSD 3-Clause License
-
-Copyright (c) 2024, 🍀☀🌕🌥 🌊
-All rights reserved.
-
-Redistribution and use in source and binary forms, with or without
-modification, are permitted provided that the following conditions are met:
-
-1. Redistributions of source code must retain the above copyright notice, this
-   list of conditions and the following disclaimer.
-
-2. Redistributions in binary form must reproduce the above copyright notice,
-   this list of conditions and the following disclaimer in the documentation
-   and/or other materials provided with the distribution.
-
-3. Neither the name of the copyright holder nor the names of its
-   contributors may be used to endorse or promote products derived from
-   this software without specific prior written permission.
-
-THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
-FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
-DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
-SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
-CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
-OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-*****************************************************************************/
-
-/**
- * @file forward_declarations.h
- * @brief Forward declarations for thread_pool module
- * 
- * This file contains forward declarations to reduce compile-time dependencies
- * and prevent circular includes within the thread_pool module.
- */
-
-namespace kcenon::thread {
-    
-    // Core types
-    class thread_pool;
-    struct thread_pool_config;
-    
-    // Worker types
-    class thread_worker;
-    enum class worker_state;
-    
-    // Builder types
-    class thread_pool_builder;
-    class pool_factory;
-    
-    // Policy types
-    struct worker_policy;
-    enum class scheduling_policy;
-    
-    // Async types (forward declare if coroutines are available)
-    #if __cplusplus >= 202002L && __has_include(<coroutine>)
-    template<typename T>
-    class task;
-    
-    template<typename T>
-    class awaitable;
-    #endif
-    
-} // namespace kcenon::thread
